% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn_test.R
\name{conn_test}
\alias{conn_test}
\title{Tests whether the website is properly connected}
\usage{
conn_test(url)
}
\arguments{
\item{url}{A website url address.}
}
\value{
A string that returns "ok" if the connection is normal 
and "no" otherwise.
}
\description{
Tests whether the website is properly connected.
}
\examples{
# Example(Not run)
# conn_test("www.baidu.com")

}
