% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig.R
\name{ig}
\alias{ig}
\title{Install GitHub packages from hub.fastgit.org}
\usage{
ig(x)
}
\arguments{
\item{x}{Repository address in the format username/repo.}
}
\description{
When you want to download an R package on Github, but you can't access Github,
this function helps you download and install the package from another website,
which is in real-time sync with Github.
}
\examples{

# Example
ig("yihui/xfun")

}
