% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig.R
\name{ig}
\alias{ig}
\title{Install R packages on Github from hub.fastgit.org}
\usage{
ig(x)
}
\arguments{
\item{x}{Repository address in the format username/repo.}
}
\description{
When you want to install an R package on GitHub, but you can't access GitHub,
this function helps you download and install the package from another website,
which is in real-time sync with GitHub.
}
\examples{

# Example
ig("yihui/xfun")

}
