% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_github.R
\name{install_github}
\alias{install_github}
\title{Install R packages on GitHub from hub.fastgit.org}
\usage{
install_github(x)
}
\arguments{
\item{x}{Repository address in the format username/repo.}
}
\description{
When you want to install an R package on GitHub, but you can't access GitHub,
this function helps you download and install the package from another website,
which is in real-time sync with GitHub.
}
\examples{

# Example
install_github("yihui/xfun")

}
